check if user in database, started middleware for auth

+++ server.ts
@@ -2,7 +2,7 @@ import { Payload } from "https://deno.land/x/djwt@v2.4/mod.ts";
 import { Application, Context, Router } from "https://deno.land/x/oak@v10.6.0/mod.ts";
 import { Client } from "https://deno.land/x/mysql@v2.10.2/mod.ts";
 import {create_jwt, generate_key, verify_jwt, hash_password} from "./auth.ts";
-import { create_db_client, insert_payments_db, delete_payments_db, get_payments_db, delete_user_db } from "./database_funcs.ts";
+import { create_db_client, insert_payments_db, delete_payments_db, get_payments_db, delete_user_db, check_user_exists } from "./database_funcs.ts";
 
 
 /**
@@ -23,6 +23,45 @@ export async function start_server(port: number, db_hostname: string, db_user: s
     const client: Client = await create_db_client(db_hostname, db_user, db_name, db_password);
 
 
+    // AUTH
+    // auth middleware to check if token was sent + check token
+    app.use(async (ctx, next)=>{
+        app.state = {user_id: undefined}; // reset state
+
+        // get data sent by frontend
+        const body = await ctx.request.body({type: "json"}).value;
+        
+        // check if token in body
+        if ("token" in body) {
+            console.log("jwt found"); // testing
+            const jwt: string = body.token as string; // extract jwt
+
+            try {
+                const payload: Payload = await verify_jwt(jwt, key); // verify jwt
+                const user_id: number = payload.user_id as number; // get user_id (claim)
+
+                // check if user exists
+                const exists: boolean = await check_user_exists(user_id, client); 
+                if (exists == false) {
+                    throw new Error(`user with id ${user_id} does not exist`);
+                }
+
+                // if jwt valid and user exists
+                app.state = {user_id: user_id};
+                
+            } catch (error) {
+
+                // if jwt invalid or user does not exist
+                ctx.response.status = 403; // Forbidden
+                ctx.response.body = error.message; // just for testing
+                return;
+            }
+        }
+
+        await next(); // go ahead (handle routes)
+    });
+
+
     // USER 
     // handle post request to /signup
     router.post("/signup", async (ctx) => handle_post_signup(ctx, key, client));
@@ -36,13 +75,13 @@ export async function start_server(port: number, db_hostname: string, db_user: s
 
     // PAYMENTS
     // handle post request to /payments
-    router.post("/payments", async (ctx) => handle_post_payments(ctx, key, client));
+    router.post("/payments", async (ctx) => handle_post_payments(ctx, client, app));
 
     // handle post request to /addpayments
-    router.post("/addpayments", async (ctx) => handle_post_add_payments(ctx, key, client));
+    router.post("/addpayments", async (ctx) => handle_post_add_payments(ctx, client, app));
 
     // handle delete request to /delpayments
-    router.delete("/delpayments", async (ctx) => handle_delete_del_payments(ctx, key, client));
+    router.delete("/delpayments", async (ctx) => handle_delete_del_payments(ctx, client, app));
 
     
     app.use(router.routes()); // use router
@@ -54,31 +93,25 @@ export async function start_server(port: number, db_hostname: string, db_user: s
  * handle_post_payments returns payments as response
  * 
  * @param ctx
- * @param key
  * @param client
  */
-async function handle_post_payments(ctx: Context, key: CryptoKey, client: Client) {
+async function handle_post_payments(ctx: Context, client: Client, app: Application) {
 
-    // verify jwt and get user_id
-    const body = await ctx.request.body({type: "json"}).value; // {token: "some_token"}
-    const jwt: string = body.token;
-
-    try {
-        const payload: Payload = await verify_jwt(jwt, key);
-        const user_id: number = (payload.user_id as number);
-        console.log("sending payments of user with id:", user_id); // just for testing
-
-        // get all payments with user_id from database
-        const payments: any[] = await get_payments_db(user_id, client);
-
-        ctx.response.status = 200; // ok
-        ctx.response.body = payments; // send as response
-
-
-    } catch (error) {
-        // handle invalid jwt
-        ctx.response.status = 403; // forbidden
+    // get user_id from application state (passed by middleware)
+    const user_id: number = app.state.user_id as number;
+    
+    // no valid user_id -> no token was sent
+    if (user_id === undefined){
+        ctx.response.status = 403; // Forbidden
+        return;
     }
+
+    // get all payments with user_id from database
+    const payments: any[] = await get_payments_db(user_id, client);
+
+    // response
+    ctx.response.status = 200; // ok
+    ctx.response.body = payments; // send as response
 }
 
 
@@ -132,7 +165,7 @@ async function handle_post_login(ctx: Context, key: CryptoKey, client: Client) {
  * @param key
  * @param client
  */
-async function handle_post_signup(ctx: Context, key:CryptoKey, client: Client) {
+async function handle_post_signup(ctx: Context, key: CryptoKey, client: Client) {
     // read body (json)
     let body = await ctx.request.body({type: "json"}).value;
 
@@ -170,43 +203,41 @@ async function handle_post_signup(ctx: Context, key:CryptoKey, client: Client) {
 /**
  * handle_post_add_payments handles new payments that need to be inserted to db
  * 
- * @param ctx 
- * @param key 
+ * @param ctx
+ * @param client
+ * @param app
  */
-async function handle_post_add_payments(ctx: Context, key: CryptoKey, client: Client) {
+async function handle_post_add_payments(ctx: Context, client: Client, app: Application) {
     // read body as json
     const body = await ctx.request.body({type: "json"}).value; // {token: some_valid_token, payments: [payment1, payment2, ...]}
-
-    const jwt = body.token; // get token
     const payments = body.payments; // payments
 
-    try {
-        const payload: Payload = await verify_jwt(jwt, key); // check jwt
-        const user_id = payload.user_id; // get user id
+    const user_id: number = app.state.user_id as number; // get user_id from apps state
 
-        // add user_id to each payment (to store in database)
-        for (let i = 0; i < payments.length; i++) {
-            payments[i].user_id = user_id;
-        }
-
-        // insert payments to database
-        const message: string = await insert_payments_db(payments, client);
-        
-        // check for errors
-        if (message === "payments added") { // "payments added" if successful
-            ctx.response.status = 200; // ok
-            ctx.response.body = message;
-        }
-        else {
-            // insertion went wrong (incorrect data, missing fields)
-            ctx.response.status = 500; // internal server error
-            ctx.response.body = message;
-        }
-
-    } catch (error) {
-        // jwt invalid
+    // no valid user_id -> no token was sent
+    if (user_id === undefined){
         ctx.response.status = 403; // Forbidden
+        return;
     }
+
+    // add user_id to each payment (to store in database)
+    for (let i = 0; i < payments.length; i++) {
+        payments[i].user_id = user_id;
+    }
+
+    // insert payments to database
+    const message: string = await insert_payments_db(payments, client);
+    
+    // check for errors
+    if (message === "payments added") { // "payments added" if successful
+        ctx.response.status = 200; // ok
+    }
+    else {
+        // insertion went wrong (incorrect data, missing fields)
+        ctx.response.status = 500; // internal server error
+    }
+
+    ctx.response.body = message;
 }
 
 
@@ -214,41 +245,34 @@ async function handle_post_add_payments(ctx: Context, key: CryptoKey, client: Cl
  * handle_delete_del_payments handles delete requests to remove payments
  * 
  * @param ctx 
- * @param key 
- * @param client
+ * @param client 
+ * @param app
  */
-async function handle_delete_del_payments(ctx: Context, key: CryptoKey, client: Client) {
+async function handle_delete_del_payments(ctx: Context, client: Client, app: Application) {
     // read body as json
     const body = await ctx.request.body({type: "json"}).value;
-
     const payment_ids = body.payment_ids; // get list of payment ids (to delete)
-    const jwt = body.token; // get jwt
 
-    // check jwt
-    try {
-        const payload: Payload = await verify_jwt(jwt, key); // check if jwt is valid
-        const user_id: number = (payload.user_id as number); // get id of user
+    const user_id: number = app.state.user_id as number; // get id of user (state)
 
-        console.log("User with id", user_id, "deletes payments:", payment_ids);
-
-
-        // delete payments (payment_ids) from database
-        const message: string = await delete_payments_db(payment_ids, user_id, client);
-
-        // error handling
-        if (message === "payments deleted") {
-            ctx.response.status = 200; // ok
-            ctx.response.body = message;
-        }
-        else {
-            ctx.response.status = 500; // Internal server error
-            ctx.response.body = message;
-        }
-
-    } catch (error) {
-        // jwt invalid
+    // no valid user_id -> no token was sent
+    if (user_id === undefined){
         ctx.response.status = 403; // Forbidden
+        return;
     }
+
+    // delete payments (payment_ids) from database
+    const message: string = await delete_payments_db(payment_ids, user_id, client);
+
+    // error handling
+    if (message === "payments deleted") {
+        ctx.response.status = 200; // ok
+    }
+    else {
+        ctx.response.status = 500; // Internal server error
+    }
+
+    ctx.response.body = message;
 }
 
 
@@ -283,7 +307,7 @@ async function handle_delete_del_user(ctx: Context, key: CryptoKey, client: Clie
         }
 
         // remove access to payments / resources on server
-        
+
     } catch (error) {
         ctx.response.status = 403; // forbidden
     }
